ggml-metal: take the FA block-skip reduction in float, not half
simd_max() over a half operand is mis-lowered at wave64 by this AMD Metal
compiler. In kernel_flash_attn_ext_vec_w64's fully-masked-block skip it can
return a value <= -MAXHALF when every lane demonstrably holds 0, so the
kernel skips blocks full of live keys and the model attends to a fraction of
its context. Characterised behaviourally: MAXHALF reads 65504, the comparison
((half)0 <= -MAXHALF) evaluates false, and the branch is taken anyway; merely
adding another read of sm[] makes it behave. Taking the reduction in float is
correct, and the same kernel already reduces floats correctly in the online
softmax.

It is silent, which is why it survived every op test. The per-key mask select
still zeroes real masked keys, so the output stays finite and plausible - just
computed over part of the cache. With no mask tensor at all it fires on every
block and the kernel returns all zeros, which is the only form the op suite
caught (ERR = inf from 0/0 on the mask=0 cases).

The tell was the performance curve, not the audio: the broken path was
perfectly flat in context, charging the same at S=812 as at S=112 when the
K+V traffic alone owes ~1 ms/token. It was not reading the cache.

Effect on the engine at N=200: rms 5391 with 0 full-scale samples and 38,705
levels, against the broken build's rms 12,979 with 13,930 full-scale samples
(1.81%). Against the eager path the fixed build reads rmsdiff 4115 - the
token-level-divergence signature, i.e. a legitimately different song from
changed summation order, which is expected here and is not expected to be
bit-identical. Against a build with the skip disabled outright it reads 2.24,
inside the noise band, so the skip is once again the no-op it is meant to be.

Context term: 23.84 -> 1.12 us per cache step, 21.3x.

Anyone touching the other wave64 kernels should grep for 16-bit simd
reductions; this is a compiler property, not a property of this kernel.

Also adds decode-shaped FLASH_ATTN_EXT coverage to test-backend-ops,
including the engine's exact permuted layout (ns10 = 1024, which the existing
permuted cases never reach because they allocate a doubled first dimension)
and nsg > 1 shapes.

+++ external/ggml/src/ggml-metal/ggml-metal.metal
@@ -8174,10 +8174,18 @@ kernel void kernel_flash_attn_ext_vec_w64(
                 sm[tiisg] = pm[ic + tiisg];
             }
 
-            // skip -INF blocks. this is also what keeps the over-allocated tail of the decode
-            // cache nearly free: a fully masked block costs one C-wide mask read and no K/V
-            // traffic at all.
-            if (simd_max(sm[tiisg]) <= -MAXHALF) {
+            // skip fully -INF blocks. this is also what keeps the over-allocated tail of the
+            // decode cache nearly free: a fully masked block costs one C-wide mask read and
+            // no K/V traffic at all.
+            //
+            // the reduction is taken in float, and that cast is load-bearing. simd_max() over
+            // a half operand is mis-lowered at wave64 by this AMD Metal compiler: it can
+            // return a value <= -MAXHALF when every lane demonstrably holds 0, which skips
+            // live blocks. it is silent - the per-key mask select still zeroes real masked
+            // keys, so the output stays finite and plausible, just computed over a fraction
+            // of the cache. with no mask tensor at all (sm[] written once by the zero-init
+            // and never again) it fires on every block and the kernel returns all zeros.
+            if (simd_max((float) sm[tiisg]) <= -(float) MAXHALF) {
                 continue;
             }
